fix: metadata editor save failure handling, VSlider, file split, onChange guard
